change service name, add logging and database healthcheck
