+++ backend-python/auth_service/Dockerfile
@@ -11,8 +11,8 @@ RUN pip install --no-cache-dir -r requirements.txt
 COPY shared /shared
 RUN pip install /shared
 
-COPY auth_service/app ./app
+COPY auth_service/auth_app ./auth_app
 
 EXPOSE 8000
 
-CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8000"]
+CMD ["uvicorn", "auth_app.main:app", "--host", "0.0.0.0", "--port", "8000"]
